requirements.txt
cycler==0.10.0
decorator==4.2.1
-e git+git@github.com:thejoeejoee/deepgaze.git@28500769862f7eac06efea9b6b5fbdecc057fdef#egg=deepgaze
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
matplotlib==2.1.2
networkx==2.1
numpy==1.14.0
opencv-python==3.4.0.12
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pkg-resources==0.0.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyWavelets==0.5.2
scikit-image==0.13.1
scipy==1.0.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
django
keras
tensorflow
mysqlclient
scp
paramiko
h5py
